2.3.4:
Referring URL check is now compatible with the `hostnames` option of `apostrophe-workflow`, ensuring that the persona prefix is accepted on the first access as long as it comes from a relevant site. If you are implementing any other kind of multisite system, you will want to override the `ourReferrer` method at project level to make the test more inclusive.